fix: long type for file size (for huge files > 2Go) && smaller buffer to avoid filling JVM RAM && feat: exec time for perfs

+++ src/Common/Fichier.java
@@ -6,6 +6,6 @@ public interface Fichier extends Serializable {
     
     String getNom();
 
-    int getSize();
+    long getSize();
 
 }
